+++ src/countries/al.ts
@@ -11,7 +11,10 @@ export const al: Country = {
 		sq: 'Shqipëria',
 		pt: 'Albânia',
 		tr: 'Arnavutluk',
+		fa: 'آلبانی',
+		uk: 'Албанія',
 		xa: 'Himni i Flamurit',
+		xt: 'Himni i Flamurit',
 	},
 	flag: '🇦🇱',
 }