+++ src/countries/ir.ts
@@ -1,0 +1,20 @@
+import { Country } from './Country'
+
+export const ir: Country = {
+	code: 'ir',
+	name: {
+		en: 'Iran',
+		ar: 'إيران',
+		de: 'Iran',
+		sv: 'Iran',
+		da: 'Iran',
+		sq: 'Irani',
+		pt: 'Irão',
+		tr: 'İran',
+		fa: 'ایران',
+		uk: 'Іран',
+		xa: 'سرود ملی جمهوری اسلامی ایران',
+		xt: 'سرود ملی جمهوری اسلامی ایران',
+	},
+	flag: '🇮🇷',
+}