+++ src/countries/us.ts
@@ -11,7 +11,10 @@ export const us: Country = {
 		sq: 'Shtetet e Bashkuara të Amerikës',
 		pt: 'Estados Unidos da América',
 		tr: 'Amerika Birleşik Devletleri',
+		fa: 'ایالات متحده آمریکا',
+		uk: 'Сполучені Штати Америки',
 		xa: 'The Star-Spangled Banner',
+		xt: 'The Star-Spangled Banner',
 	},
 	flag: '🇺🇸',
 }